dev-requirements.txt
sphinx >= 4, != 5.2.0, < 6
pydata-sphinx-theme
pytest
pytest-coverage
numpy
matplotlib
pillow
seaborn
statsmodels
joblib
importlib-metadata <5.0
flake8
check-manifest
plotly
absl-py
graphviz
packaging
jupyterlite-sphinx
sphinxcontrib-svg2pdfconverter[CairoSVG]
